Clean-Up #3314
