+++ js/src/ui/Form/TypedInput/typedInput.css
@@ -15,19 +15,17 @@
 /* along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-.center {
-  text-align: center;
-}
+.inputs {
+  padding-top: 2px;
+  overflow-x: hidden;
 
-.address {
-  vertical-align: top;
-  display: inline-block;
-}
-
-.identityicon {
-  margin: -8px 0.5em;
-}
-
-.funcparams {
-  padding-left: 3em;
+  label {
+    line-height: 22px;
+    pointer-events: none;
+    color: rgba(255, 255, 255, 0.498039);
+    -webkit-user-select: none;
+    font-size: 12px;
+    top: 11px;
+    position: relative;
+  }
 }
